fix: move from to network utils
