+++ src/adapters/network.js
@@ -316,15 +316,7 @@ const handleHttpRequest = async (requestType = 'post', ...httpArgs) => {
   return { httpResponse, processedResponse };
 };
 
-function getCompatibleStatusCode(status) {
-  if (status === HttpStatus.NOT_FOUND) {
-    return HttpStatus.BAD_REQUEST;
-  }
-  if (isHttpStatusSuccess(status)) {
-    return HttpStatus.OK;
-  }
-  return status;
-}
+
 
 module.exports = {
   httpSend,
@@ -338,5 +330,4 @@ module.exports = {
   getPayloadData,
   getFormData,
   handleHttpRequest,
-  getCompatibleStatusCode,
 };
